Add support for recent Institute of Physics (IOP) journals articles
Multiple journals are supported like this, as IOP hosts all of them
on the same website. I Have some examples of working articles [1], [2].
The scraping matches current formatting. Unfortunately the formatting
was different a few years back, for example [3]. Those are not handled
by these rules, and would need to implement separately.

[1]: http://iopscience.iop.org/1367-2630/16/6/063043/article
[2]: http://iopscience.iop.org/0029-5515/54/6/064001/article
[3]: http://iopscience.iop.org/1367-2630/8/9/188/fulltext/